parsed ld.py into functions+script, tests for ld added, script for SuSiE added
